try catch improvments

+++ application/lib/Controllers/PostHierarchyController.php
@@ -23,33 +23,31 @@ class PostHierarchyController
     {
         // https://github.com/php-fig/http-message/blob/master/src/ServerRequestInterface.php#L167
         try {
-			$data = $request->getParsedBody();			
+			$data = $request->getParsedBody();
+			if(!$data)
+				throw new \InvalidArgumentException("Invalid input (cannot be decoded)");
             $tree = new Tree($data);
-			return $response->withJson($tree, 200);
-
             $this->dbConnection->beginTransaction();
-            try {
-                $this->dbConnection->exec('DELETE FROM employees');
-
-                foreach ($tree as $child => $parent) {
-                    $sql = sprintf(
-                        "INSERT INTO employees (name, supervisor) VALUES ('%s','%s') ",
-                        $child,
-                        $parent
-                    );
-
-                    $this->dbConnection->exec($sql);
-                }
-
-                $this->dbConnection->commit();
-            } catch (\PDOException $exception) {
-                $this->dbConnection->rollBack();
-                throw $exception;
-            }
+			$this->dbConnection->exec('DELETE FROM employees');
+			foreach ($tree as $child => $parent) {
+				$sql = sprintf("INSERT INTO employees (name, supervisor) VALUES ('%s','%s') ",
+					$child,
+					$parent
+				);
+				$this->dbConnection->exec($sql);
+			}
+			$this->dbConnection->commit();
 
             return $response->withJson($tree, 200, JSON_PRETTY_PRINT);
-        } catch (\InvalidArgumentException $e) {
-            return $response->withJson(['error' => $e->getMessage()], 422);
-        }
+        } catch (\PDOException $exception) {
+			$this->dbConnection->rollBack();
+			return $response->withJson(['error' => $e->getMessage()], 500);
+		} catch(\Personia\Tree\RecursionException $e) {
+			return $response->withJson(['error' => $e->getMessage()], 422);
+		} catch (\InvalidArgumentException $e) {
+            return $response->withJson(['error' => $e->getMessage()], 400);
+        }catch (\TypeError $e) {
+			return $response->withJson(['error' => $e->getMessage()], 400);
+		}
     }
 }
